+++ src/modifiers.rs
@@ -101,7 +101,7 @@ impl InitModifier for PositionCircleModifier {
     // Spawn random point on/in circle
     let theta = rand() * tau;
     let dir = tangent * cos(theta) + bitangent * sin(theta);
-    ret.pos = c + r * dir;
+    ret.pos = c + r * dir + transform[3].xyz;
     // Velocity away from center
     ret.vel = dir * speed;
     // <<< [PositionCircleModifier]
@@ -162,7 +162,7 @@ impl InitModifier for PositionSphereModifier {
     var x = sinphi * cos(theta);
     var y = sinphi * sin(theta);
     var dir = vec3<f32>(x, y, z);
-    ret.pos = c + r * dir;
+    ret.pos = c + r * dir + transform[3].xyz;
     // Radial velocity away from sphere center
     ret.vel = dir * speed;
     // <<< [PositionSphereModifier]
@@ -170,6 +170,101 @@ impl InitModifier for PositionSphereModifier {
             self.center.to_wgsl_string(),
             radius_code,
             self.speed.to_wgsl_string()
+        );
+    }
+}
+
+
+/// An initialization modifier spawning particles inside a truncated 3D cone.
+///
+/// The 3D cone is oriented along the Y axis, with its origin at the center of the top
+/// circle truncating the cone. The center of the base circle of the cone is located at
+/// a positive Y.
+///
+/// Particles are spawned somewhere inside the volume or on the surface of a truncated
+/// 3D cone defined by its base radius, its top radius, and the height of the cone section.
+///
+/// The particle velocity is initialized to a random speed along the direction going from
+/// the cone apex to the particle position.
+#[derive(Default, Clone, Copy)]
+pub struct PositionCone3dModifier {
+    /// The cone height along its axis, between the base and top radii.
+    pub height: f32,
+    /// The cone radius at its base, perpendicularly to its axis.
+    pub base_radius: f32,
+    /// The cone radius at its truncated top, perpendicularly to its axis.
+    /// This can be set to zero to get a non-truncated cone.
+    pub top_radius: f32,
+    /// The speed of the particles on spawn.
+    pub speed: Value<f32>,
+    /// The shape dimension to spawn from.
+    pub dimension: ShapeDimension,
+}
+
+impl InitModifier for PositionCone3dModifier {
+    fn apply(&self, init_layout: &mut InitLayout) {
+        if matches!(self.dimension, ShapeDimension::Surface) {
+            unimplemented!("TODO");
+        }
+        // let radius_code = match self.dimension {
+        //     ShapeDimension::Surface => {
+        //         // Constant radius
+        //         format!("let r = {};", self.radius.to_wgsl_string())
+        //     }
+        //     ShapeDimension::Volume => {
+        //         // Radius uniformly distributed in [0:1], then scaled by ^(1/3) in 3D
+        //         // to account for the increased surface covered by increased radii.
+        //         // https://stackoverflow.com/questions/54544971/how-to-generate-uniform-random-points-inside-d-dimension-ball-sphere
+        //         format!(
+        //             "var r = pow(rand(), 1./3.) * {};",
+        //             self.radius.to_wgsl_string()
+        //         )
+        //     }
+        // };
+        init_layout.position_code = format!(
+            r##"
+    // >>> [PositionCone3dModifier]
+    // Truncated cone height
+    let h0 = {0};
+    // Random height ratio
+    let alpha_h = pow(rand(), 1.0 / 3.0);
+    // Random delta height from top
+    let h = h0 * alpha_h;
+    // Top radius
+    let rt = {1};
+    // Bottom radius
+    let rb = {2};
+    // Radius at height h
+    let r0 = rt + (rb - rt) * alpha_h;
+    // Random delta radius
+    let alpha_r = sqrt(rand());
+    // Random radius at height h
+    let r = r0 * alpha_r;
+    // Random base angle
+    let theta = rand() * tau;
+    let cost = cos(theta);
+    let sint = sin(theta);
+    // Random position relative to truncated cone origin (not apex)
+    let x = r * cost;
+    let y = h;
+    let z = r * sint;
+    let p = vec3<f32>(x, y, z);
+    let p2 = transform * vec4<f32>(p, 1.0);
+    ret.pos = p2.xyz;
+    // Emit direction
+    let rb2 = rb * alpha_r;
+    let pb = vec3<f32>(rb2 * cost, h0, rb2 * sint);
+    let dir = transform * vec4<f32>(normalize(pb - p), 0.0);
+    // Emit speed
+    let speed = {3};
+    // Velocity away from cone top/apex
+    ret.vel = dir.xyz * speed;
+    // <<< [PositionCone3dModifier]
+"##,
+            self.height.to_wgsl_string(),
+            self.top_radius.to_wgsl_string(),
+            self.base_radius.to_wgsl_string(),
+            self.speed.to_wgsl_string(),
         );
     }
 }